add icon weather type

+++ app/src/main/java/co/develhope/meteoapp/data/domainmodel/RowForecast.kt
@@ -6,6 +6,6 @@ import org.threeten.bp.OffsetDateTime
 data class RowForecast(
     val time: OffsetDateTime,
     val weatherCondition: WeatherCondition,
-    val humidity: Int,
-    val temp: Int
+    val humidity: Double,
+    val temp: Double
 )
